libs/hr_parser/requirements.txt
click>=8.0.3
environs>=10.3.0
fastapi>=0.104.1
langchain>=0.1.8
openai<1
markdownify>=0.11.6
pdfminer-six>=20231228
python-multipart>=0.0.9
mangum>=0.17.0
uvicorn>=0.27.1
PyMuPDF>=1.23.17
pydantic>=1.10.0
Pillow>=10.2.0
opencv-python>=4.9.0.80
numpy>=1.26.4
